backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
pydantic[email]
pydantic-settings
python-dotenv
alembic
bcrypt
python-jose
aiosmtplib
slowapi
python-socketio
cryptography
python-multipart
httpx
# Testing
pytest
pytest-asyncio
aiosqlite
aiohttp
cloudinary
# 
google-genai
Pillow
